move bottom tabs navigator to own file
