+++ src/lib/Scenes/BottomTabs/BottomTabsNavigator.tsx
@@ -1,0 +1,96 @@
+import { AppModule } from "lib/AppRegistry"
+import { NativeViewController } from "lib/Components/NativeViewController"
+import { AppStore } from "lib/store/AppStore"
+import { useScreenDimensions } from "lib/utils/useScreenDimensions"
+import React, { useEffect, useRef } from "react"
+import { Animated, View } from "react-native"
+import { BottomTabs } from "./BottomTabs"
+import { BottomTabType } from "./BottomTabType"
+
+const NavStack = ({
+  tabName,
+  rootModuleName,
+  rootModuleProps,
+}: {
+  tabName: BottomTabType
+  rootModuleName: AppModule
+  rootModuleProps: object
+}) => {
+  return (
+    <NativeViewController
+      viewName="TabNavigationStack"
+      viewProps={{
+        tabName,
+        rootModuleName,
+        rootModuleProps,
+      }}
+    />
+  )
+}
+
+export const BottomTabsNavigator = () => {
+  const selectedTab = AppStore.useAppState((state) => state.bottomTabs.selectedTab || "home")
+  const tabProps = AppStore.useAppState((state) => state.bottomTabs.sessionState.tabProps)
+  const { bottom } = useScreenDimensions().safeAreaInsets
+  return (
+    <View style={{ flex: 1, paddingBottom: bottom }}>
+      <FadeBetween
+        views={[
+          <NavStack tabName="home" rootModuleName="Home" rootModuleProps={tabProps.home ?? {}} />,
+          <NavStack tabName="search" rootModuleName="Search" rootModuleProps={tabProps.search ?? {}} />,
+          <NavStack tabName="inbox" rootModuleName="Inbox" rootModuleProps={tabProps.inbox ?? {}} />,
+          <NavStack tabName="sell" rootModuleName="SellTabApp" rootModuleProps={tabProps.sell ?? {}} />,
+          <NavStack tabName="profile" rootModuleName="MyProfile" rootModuleProps={tabProps.profile ?? {}} />,
+        ]}
+        activeIndex={["home", "search", "inbox", "sell", "profile"].indexOf(selectedTab)}
+      />
+      <BottomTabs />
+    </View>
+  )
+}
+
+const FadeBetween: React.FC<{ views: JSX.Element[]; activeIndex: number }> = ({ views, activeIndex }) => {
+  const opacities = useRef(views.map((_, index) => new Animated.Value(index === activeIndex ? 1 : 0))).current
+  const lastActiveIndex = usePrevious(activeIndex)
+  useEffect(() => {
+    if (lastActiveIndex < activeIndex) {
+      // fade in screen above, then make previous screen transparent
+      Animated.spring(opacities[activeIndex], { toValue: 1, useNativeDriver: true, speed: 100 }).start(() => {
+        opacities[lastActiveIndex].setValue(0)
+      })
+    } else if (lastActiveIndex > activeIndex) {
+      // make next screen opaque, then fade out screen above
+      opacities[activeIndex].setValue(1)
+      requestAnimationFrame(() => {
+        Animated.spring(opacities[lastActiveIndex], { toValue: 0, useNativeDriver: true, speed: 100 }).start()
+      })
+    }
+  }, [activeIndex])
+  return (
+    <View style={{ flex: 1, overflow: "hidden" }}>
+      {views.map((v, index) => {
+        return (
+          (index === activeIndex || index === lastActiveIndex) && (
+            <View
+              key={index}
+              pointerEvents={index === activeIndex ? undefined : "none"}
+              style={{ position: "absolute", width: "100%", height: "100%" }}
+            >
+              <Animated.View style={{ opacity: opacities[index], flex: 1, backgroundColor: "white" }}>
+                {v}
+              </Animated.View>
+            </View>
+          )
+        )
+      })}
+    </View>
+  )
+}
+
+function usePrevious<T>(value: T) {
+  const ref = useRef<T>(value)
+  useEffect(() => {
+    ref.current = value
+  }, [value])
+  return ref.current
+}
